requirements.txt
async-timeout==4.0.2
aerich==0.7.1
aiomysql==0.1.1
aiosqlite==0.17.0
amqp==5.1.1
anyio==3.6.1
async-timeout==4.0.2
bcrypt==4.0.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
dictdiffer==0.9.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.85.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
kombu==5.2.4
librabbitmq==2.0.0
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
Pillow==9.2.0
prompt-toolkit==3.0.31
pyasn1==0.4.8
pydantic==1.10.2
PyMySQL==1.0.2
pyparsing==3.0.9
pypika-tortoise==0.1.6
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.2.1
PyYAML==6.0
redis==4.4.0
reportlab==3.6.12
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tomlkit==0.11.5
tortoise-orm==0.19.2
typing-extensions==4.3.0
uvicorn==0.18.3
vine==5.0.0
watchfiles==0.17.0
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1